autre update

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\EventController;
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\PlaceController;
+use App\Http\Controllers\PayementController;
 use App\Http\Controllers\SocialmediaController;
 
 
@@ -18,40 +19,32 @@ Route::get('/login', [AuthController::class, 'showformlogin'])->name('login');
 Route::post('/register', [AuthController::class, 'register'])->name('register.post');
 Route::post('/login', [AuthController::class, 'login'])->name('login.post');
 Route::get('/logout', [AuthController::class, 'logout'])->name('logout');
+// Routes pour evenements dans page d'accueil
+Route::get('/', [EventController::class, 'list_events_accueil'])->name('events.home');
+Route::get('/event/{id}', [EventController::class, 'show_event_accueil'])->name('accueil.event.show');
 
-//Route de paiement
-Route::get('/pay', [PayementController::class, 'showPaymentForm'])->name('pay.form');
-Route::post('/pay', [PayementController::class, 'processTransaction'])->name('pay.process');
+
 
 
 // Routes des catégories dans dashboard admin 
-Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
-Route::post('/categories', [CategoryController::class, 'store'])->name('categories.store');
-Route::get('/categories/edit/{id}', [CategoryController::class, 'edit'])->name('categories.edit');
-Route::put('/categories/{id}', [CategoryController::class, 'update'])->name('categories.update');
-Route::delete('/categories/{id}', [CategoryController::class, 'destroy'])->name('categories.destroy');
-Route::get('events', [CategoryController::class, 'categoryonevents'])->name('categories.events');
-
-
+Route::get('/categories', [CategoryController::class, 'index'])->name('admin.categories.index');
+Route::post('/categories', [CategoryController::class, 'store'])->name('admin.categories.store');
+Route::get('/categories/edit/{id}', [CategoryController::class, 'edit'])->name('admin.categories.edit');
+Route::put('/categories/{id}', [CategoryController::class, 'update'])->name('admin.categories.update');
+Route::delete('/categories/{id}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
+Route::get('events', [CategoryController::class, 'categoryonevents'])->name('admin.categories.events');
 // Routes des utilisateurs dans dashboard admin 
 Route::get('/users', [UserController::class, 'index'])->name('users.index');
 Route::post('/users', [UserController::class, 'store'])->name('users.store');
 Route::get('/users/edit/{id}', [UserController::class, 'edit'])->name('users.edit');
 Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
 Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
-
-// Routes pour evenements dans page d'accueil
-Route::get('/', [EventController::class, 'list_events_accueil'])->name('events.home');
-Route::get('/event/{id}', [EventController::class, 'show_event_accueil'])->name('accueil.event.show');
-
 // Routes des evenements dans dashboard admin 
 Route::post('/admin/events', [EventController::class, 'store_event_admin'])->name('admin.events.store');
 Route::get('/admin/events', [EventController::class, 'list_events_admin'])->name('admin.events.index');
 Route::get('/admin/events/edit/{id}', [EventController::class, 'edit'])->name('admin.events.edit');
 Route::put('/admin/events/{id}', [EventController::class, 'update'])->name('admin.events.update');
 Route::delete('/admin/events/{id}', [EventController::class, 'destroy'])->name('admin.events.destroy');
-
-
 // Routes des lieux dans dashboard admin 
 Route::get('/places', [PlaceController::class, 'index'])->name('admin.places.index');
 Route::post('/places', [PlaceController::class, 'store'])->name('admin.places.store');
@@ -59,17 +52,15 @@ Route::get('/places/edit/{id}', [PlaceController::class, 'edit'])->name('admin.p
 Route::put('/places/{id}', [PlaceController::class, 'update'])->name('admin.places.update');
 Route::delete('/places/{id}', [PlaceController::class, 'destroy'])->name('admin.places.destroy');
 
+
+
 // Routes des informations de l'organisateur dans dashboard organisateur
 Route::get('/organizerinfo', [UserController::class, 'mesinformation'])->name('organizer.information');
 Route::put('/organizerinfo/{id}', [UserController::class, 'updateinfo'])->name('organizer.update');
-
-
-
 Route::get('/organizer', [EventController::class, 'countevents'])->name('organizer');
 Route::get('/billes', function () {return view('dashboard.organizer.billets-vendus');})->name('organizer.billets');
 Route::get('/statistiques', function () {return view('dashboard.organizer.statistiques');})->name('organizer.statistiques');
 Route::get('/payements', function () {return view('dashboard.organizer.payements');})->name('organizer.payements');
-
 //Routes pour evenements dans dashboard organisateur
 Route::get('/organizer/events', [EventController::class, 'list_events_organizer'])->name('organizer.events.index');
 Route::get('/organizer/events/create', [EventController::class, 'create_event_organizer'])->name('organizer.events.create');
